fixed more notes, added new methods
