+++ src/main/java/com/epam/homelibrary/Main.java
@@ -3,10 +3,6 @@ package com.epam.homelibrary;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
 /**
  * Home library
  * <p>
@@ -17,7 +13,7 @@ public class Main {
     public static final Logger logger = LogManager.getLogger(Main.class);
 
     public static void main(String[] args) {
-        Controller controller = new Controller();
-        controller.operate();
+        LibraryAPI libraryAPI = new LibraryAPI();
+        libraryAPI.operate();
     }
 }
